api/requirements.txt
Flask==3.0.3
flask-smorest==0.45.0
Flask-Limiter==3.7.0
Flask-Caching==2.3.0
gunicorn==22.0.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
marshmallow==3.21.2
structlog==24.1.0
python-dotenv==1.0.1
redis==5.0.4
flask-cors==4.0.1
